删除checkutype+移动其它
